Sort out .d modules declarations
