--- src/models/d2.d.ts
@@ -1,3 +1,0 @@
-declare module "d2/uid" {
-    export function generateUid(): string;
-}
